requirements.txt
absl-py==0.12.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
conllu==4.4
cycler==0.10.0
datasets==1.5.0
decorator==5.0.6
deepspeed==0.3.13
defusedxml==0.7.1
dill==0.3.3
elasticsearch==7.12.0
entrypoints==0.3
faiss-cpu==1.7.0
filelock==3.0.12
fire==0.4.0
fsspec==0.9.0
future==0.18.2
git-python==1.0.3
gitdb==4.0.7
gitpython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
grpcio==1.37.0
huggingface-hub==0.0.8
idna==2.10
importlib-metadata==3.10.0
importlib-resources==5.1.2
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jinja2==2.11.3
joblib==1.0.1
jsonlines==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
markdown==3.3.4
markupsafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
multiprocess==0.70.11.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
ninja==1.10.0.post2
nltk==3.6.1
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.2.0
pip==21.0.1
pluggy==0.13.1
portalocker==2.0.0
prometheus-client==0.10.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.7
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.2
pydeck==0.6.1
pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.7.2
sacrebleu==1.5.1
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.2
send2trash==1.5.0
sentencepiece==0.1.95
seqeval==1.2.2
setuptools==52.0.0
six==1.15.0
smmap==4.0.0
streamlit==0.79.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardx==1.8
tensorflow-datasets==4.2.0
tensorflow-metadata==0.29.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
toolz==0.11.1
# torch==1.8.1+cu111
torchaudio==0.8.1
# torchvision==0.9.1+cu111
torchvision==0.9.1
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
transformers==4.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
xxhash==2.0.0
zipp==3.4.1
